Add files via upload
The purpose of this project was to find an accurate information of Booking , travelling . 
This is for the customer who are facing issues or various difficulties while booking their ticket by visiting to the reservation counter or agent .
In this we have use PHP and Mysql , In this We have used the PHP version 7 , Mysql Version 5.6 , and configure this project on Xampp Server and for front end We use Html, CSS , Java-script.

This is a mini project for single user/login only for customers to fulfill their requirements. By this they can book their ticket easily & directly.The purpose was to find an accurate information of Booking , travelling . This is for the customer who are facing issues or various difficulties while booking their ticket by visiting to the reservation counter or agent . In this we have use PHP and Mysql , In this We have used the PHP version 7 , Mysql Version 5.6 , and configure this project on Xampp Server and for front end We use Html, CSS , Java-script. This is a mini project for single user/login only for customers to fulfill their requirements. By this they can book their ticket easily & directly.
Other creators